address compile warnings

+++ src/Memory.cpp
@@ -166,7 +166,7 @@ void Memory::attachToCore(libretro::Core* core, int consoleId)
   }
 
   bool hasValidRegion = false;
-  for (int i = 0; i < g_memoryRegionCount; i++)
+  for (size_t i = 0; i < g_memoryRegionCount; i++)
   {
     if (g_memoryRegionData[i] != NULL)
     {
